ARM: tegra: medcom-wide: Remove extra panel power supply
Simple panels can only have a single power supply. The second listed
supply is not needed because it is also the input supply of the first
supply and therefore will always be on at the same time.

In retrospect the panel probably doesn't qualify as simple since it
apparently does need both of these supplies, even if in the case of the
Medcom Wide it isn't necessary to explicitly hook them up.

Signed-off-by: Thierry Reding <treding@nvidia.com>

[ upstream commit: 90b1307ac545e6e7e26f7a6be9e9ef49c4a650b1 ]