dt-bindings: i2c-gpio: Use 'deprecated' keyword on deprecated properties
The i2c-gpio binding is the only case of 'not: true' being used for a
deprecated property. Update it to use the json-schema 2019.09
'deprecated' keyword instead. Define the type too in order to keep the
meta-schema happy.

This will disable warnings for these properties until the dtschema tools
are updated to handle 'deprecated', but they are only used in a few at91
dts files.

Cc: Wolfram Sang <wolfram@the-dreams.de>
Cc: linux-i2c@vger.kernel.org
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: a451c3be0fa1a4829a7597e5611126c041a4b778 ]