ARM: dts: imx25-pinfunc: Fix GPT function names
According to the reference manual of the i.MX25 the function name for
the GPT capture input is called CAPIN1. So fix the names to follow the
same naming scheme as used for the other pads and functions. Also add
one missing pad function.

There are no mainline users that needs adaption.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 0b784a7b1c53c3989e9ee2dfdd3fc9b6b14b5256 ]