ARM: dts: imx6q-icore-ofcap10: Use 10.1" Ampire panel compatible
Adding display timings directly on device tree files make it difficult
to maintain as a same copy of timings may exist on different files or
panel-simple driver.

We have a panel-simple driver for this particular usage so supporting
on this driver will help to use the same timings on any device tree
files if the board mounted on a similar vendor display.

Engicam C.TOUCH OF 10.1" LCD board uses Ampire 10.1" TFT LCD and
it has supported by panel-simple already, so simply use that binding.

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: faa3b6dfd2527eace517e0cd1b147154c780d582 ]