ARM: dts: exynos: Override thermal by label in Midas
Using full paths to extend or override a device tree node is error prone
since if there was a typo error, a new node will be created instead of
extending the node as it was desired.  This will lead to run-time errors
that could be hard to detect.

A mistyped label on the other hand, will cause a dtc compile error
(during build time).

Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>
Link: https://lore.kernel.org/r/20200830135200.24304-26-krzk@kernel.org

[ upstream commit: e4cf0a4dba7f0bd579dc96f24d8e6134995005c5 ]