dt-bindings: interconnect: Document the support of optional path tag
Let's document that we now support specifying path tag information in the
arg cells of the 'interconnects' DT property. This information would be
populated when the xlate_extended() callback is used.

Specifying the tag in DT will allow the interconnect framework to do the
aggregation based on the tag automatically. The users can still use the
icc_set_tag() API if/when needed.

Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20200903133134.17201-3-georgi.djakov@linaro.org
Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>

[ upstream commit: 9a34e7ad2e123b4ed02a1c579939aadf148683c5 ]