ARM: dts: exynos: override GPIO keys node by label in Exynos4412 Odroid family
Using full paths to extend or override a device tree node is error
prone.  If there was a typo error, a new node will be created instead of
extending the existing node.  This will lead to run-time errors that
could be hard to detect.

A mistyped label on the other hand, will cause a dtc compile error
(during build time).

Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>
Link: https://lore.kernel.org/r/20201027170947.132725-4-krzk@kernel.org

[ upstream commit: 43552e6d5a6109df05ea038f512a20b45ca5010c ]