dt-bindings:iio:light:st,vl6180: txt to yaml format conversion.
I'm not sure anyone would use this part primarily as an ALS,
given the time of flight laser also present, but I'll stick with the
original decision on where to put the binding.

Added interrupts property as the device has a GPIO interrupt even
if the driver is not currently using it.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Acked-by: Manivannan Sadhasivam <manivannanece23@gmail.com>
Cc: Manivannan Sadhasivam <manivannanece23@gmail.com>
Cc: Peter Meerwald-Stadler <pmeerw@pmeerw.net>
Link: https://lore.kernel.org/r/20201031184854.745828-23-jic23@kernel.org

[ upstream commit: 2d948d057558cea0ebec82f4cf6e6dd1fa1c6d65 ]