dt-bindings:iio:adc:renesas,rcar-gyroadc: txt to yaml conversion.
This is a somewhat unusual device, in that it effectively does
spi offload.   That means that it doesn't act as a full SPI
master, but supports some functionality.  As such it supports
a subset of specific SPI ADCs.  There is potential for a future
clash in bindings, but as these are simple devices hopefully that
will not occur.

One addition to this from testing it against existing dts files
was to add a resets property.
This is specified in arch/arm/boot/dts/r8a7791.dtsi
If it's the dtsi that is wrong and not the binding doc, then
we can fix that instead.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Cc: Marek Vasut <marek.vasut+renesas@gmail.com>
Link: https://lore.kernel.org/r/20201031184854.745828-31-jic23@kernel.org

[ upstream commit: 8c41245872e206ec932c24d1fd63f8ff054e9855 ]