dt-bindings: memory: tegra20: emc: Document nvidia, memory-controller property
Tegra20 External Memory Controller talks to DRAM chips and it needs to be
reprogrammed when memory frequency changes. Tegra Memory Controller sits
behind EMC and these controllers are tightly coupled. This patch adds the
new phandle property which allows to properly express connection of EMC
and MC hardware in a device-tree, it also put the Tegra20 EMC binding on
par with Tegra30+ EMC bindings, which is handy to have.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>
Acked-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20201104164923.21238-6-digetx@gmail.com
Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>

[ upstream commit: e51a59f079c546b8b5c31604ad98e86b06ec93e9 ]