dt-bindings: arm: sunxi: Add SL631 with IMX179 bindings
This adds documentation for the compatible strings of the
SL631 Action Camera with IMX179.

Note that the device is sold under various different names, such as the
SJCAM SJ4000 Air or F60 Action Camera. This is a similar situation to
the Q8 tablets and just like them, the allwinner vendor is used as
fallback.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20201206165131.1041983-5-contact@paulk.fr

[ upstream commit: 46ad18e7d0e294bc2a225e40ad98fa27eb20b86d ]