arm64: dts: renesas: beacon: Fix RGB Display PWM Backlight
The backlight didn't really work correctly due to some updates that were
made in hardware.  It should be safe to apply these, because the older
hardware was never shipped to anyone, so it shouldn't break anything.
Because the display driver refers to the display as DPI, this also
renames the backlight to use DPI for consistency.

Signed-off-by: Adam Ford <aford173@gmail.com>
Link: https://lore.kernel.org/r/20201213183759.223246-10-aford173@gmail.com
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>

[ upstream commit: 7f1cb1579808f97bd211b4ad16c82f2dc191c709 ]